improve `easy_import.py`: removing inconsistent class names + references to overridden subpackages
